+++ src/lazybull/ml/__init__.py
@@ -1,6 +1,7 @@
 """机器学习模块"""
 
 from .model_registry import ModelRegistry
+from .ensemble import EnsembleModel
 from .eval_utils import (
     compute_daily_rankic,
     compute_daily_topk_returns,
@@ -16,6 +17,7 @@ from .train_core import build_rank_sample_weights, split_train_val_by_date
 
 __all__ = [
     "ModelRegistry",
+    "EnsembleModel",
     "compute_daily_rankic",
     "compute_daily_topk_returns",
     "evaluate_predictions_by_date",